# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Catalog of cat breeds",
    "version": "3.0",
    "contact": {
      "name": "API Support",
      "email": "Antonela.Velickov@fer.hr"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8010/api"
    }
  ],
  "paths": {
    "/v2/catbreeds": {
      "get": {
        "description": "Returns a collection of cat breeds",
        "responses": {
          "200": {
            "description": "Fetch successfull"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Adds new cat breed to the database",
        "requestBody": {
          "description": "Create a new breed",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "weight": {
                    "type": "string"
                  },
                  "lifespan": {
                    "type": "string"
                  },
                  "desc": {
                    "type": "string"
                  },
                  "coo": {
                    "type": "string"
                  },
                  "ct": {
                    "type": "string"
                  },
                  "eyecolors": {
                    "type": "string"
                  },
                  "earhape": {
                    "type": "string"
                  },
                  "eyeshape": {
                    "type": "string"
                  },
                  "coats": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "weight",
                  "lifespan",
                  "desc",
                  "coo",
                  "ct",
                  "eyecolors",
                  "earshape",
                  "eyeshape",
                  "coats"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Create successfull"
          },
          "400": {
            "description": "One or more of given body parameters are incorrect"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}